refacto to ease auth resources identification
