+++ src/routes/__init__.py
@@ -1,0 +1,4 @@
+from .auth import auth_routes
+from .store import basic_store
+
+__all__ = ["auth_routes", "basic_store"]
