+++ src/schemas/__init__.py
@@ -1,0 +1,1 @@
+"""Schemas package reserved for request/response validation objects."""
