+++ src/services/__init__.py
@@ -1,0 +1,1 @@
+"""Service layer for business and authentication logic."""
