+++ src/services/auth/__init__.py
@@ -1,0 +1,9 @@
+from .common import require_authenticated_user
+from .oidc import apply_logout_cookies, build_logout_redirect, is_oidc_auth_enabled
+
+__all__ = [
+    "apply_logout_cookies",
+    "build_logout_redirect",
+    "is_oidc_auth_enabled",
+    "require_authenticated_user",
+]
