+++ src/services/auth/georchestra.py
@@ -1,0 +1,5 @@
+"""geOrchestra-specific auth service hooks.
+
+Currently geOrchestra access is based on upstream headers only and does not
+require OIDC-style session/logout handling in the application layer.
+"""
