+++ src/services/register_service.py
@@ -1,0 +1,3 @@
+from ..utils.register_utils import ConfigRegister, from_xml_path
+
+__all__ = ["ConfigRegister", "from_xml_path"]
