+++ src/utils/auth/__init__.py
@@ -1,0 +1,1 @@
+"""Authentication provider helpers."""
